moved helper fn from client-sdk-base to common-sdk-base
