+++ packages/client/base/src/consts/index.ts
@@ -1,7 +1,3 @@
 export * from "./embed";
 
 export const DATADOG_CLIENT_TOKEN = "pub035be8a594b35be1887b6ba76c4029ca";
-
-export const CROSSMINT_DEV_URL = "http://localhost:3000/";
-export const CROSSMINT_STG_URL = "https://staging.crossmint.com/";
-export const CROSSMINT_PROD_URL = "https://www.crossmint.com/";
