--- packages/client/base/src/services/api/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from "./validateApiKeyAndGetCrossmintBaseUrl";
