+++ packages/client/base/src/services/api/index.ts
@@ -1,3 +1,2 @@
 export * from "./APIErrorService";
 export * from "./BaseCrossmintService";
-export * from "./helpers";
