+++ packages/common/base/src/apiKey/consts.ts
@@ -1,2 +1,6 @@
 export const CROSSMINT_API_KEY_SIGNER_PUBLICKEY_STAGING = "3hSfN4dWSwgCg1uf2yytBtK6KxK3ySFKasd2h9J2vSK5";
 export const CROSSMINT_API_KEY_SIGNER_PUBLICKEY_PROD = "8erZh8YApGck3iUSUHqATBxqMTM1Ukp9mHmvGgUWHtkK";
+
+export const CROSSMINT_DEV_URL = "http://localhost:3000/";
+export const CROSSMINT_STG_URL = "https://staging.crossmint.com/";
+export const CROSSMINT_PROD_URL = "https://www.crossmint.com/";
