+++ packages/common/base/src/apiKey/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./environmentToCrossmintBaseURL";
 export * from "./environmentToExpectedPublicKey";
 export * from "./usageOriginToPrefix";
+export * from "./validateApiKeyAndGetCrossmintBaseUrl";
